hash table impl(removing my main.c)

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -10,7 +10,7 @@
  */
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	int x;
+	unsigned long int x;
 	struct hash_table_s *TABLE = NULL;
 
 	if (size <= 0)
